requirements.txt
gradio
faiss-cpu
numpy
langchain
langchain_community
langchain_anthropic
PyPDF2
sentence-transformers
langchain_huggingface
anthropic